research/requirements.txt
openai>=1.58.0
pydantic>=2.0.0
python-dotenv>=1.0.0
